+++ src/seq_utils.cpp
@@ -344,7 +344,7 @@ void write_nexus_alignment (std::vector<Sequence>& seqs, const bool& uppercase,
 void create_vector_seq_codon_state_reconstructor (std::vector<Sequence>& origseqs,
         std::vector<Sequence>& sr_seqs, int site, std::map<std::string,
         std::vector<int> >& codon_pos) {
-    unsigned int start = site * 3;
+    unsigned int start = static_cast<unsigned int>(site) * 3u;
     for (unsigned int i = 0; i < origseqs.size(); i++) {
         std::string codon = origseqs[i].get_sequence().substr(start, 3);
         std::string setsq;
@@ -367,7 +367,7 @@ void create_vector_seq_codon_state_reconstructor (std::vector<Sequence>& origseq
 void create_vector_seq_codon_state_reconstructor_all_site (std::vector<Sequence>& origseqs,
         std::vector<Sequence>& sr_seqs, int site, std::map<std::string,
         std::vector<int> >& codon_pos) {
-    unsigned int start = site * 3;
+    unsigned int start = static_cast<unsigned int>(site) * 3u;
     for (unsigned int i = 0; i < origseqs.size(); i++) {
         std::string codon = origseqs[i].get_sequence().substr(start, 3);
         std::string setsq(61, '0');
